// ClientApp/app/Models/LinkSetUps/linkSetUp.ts
export class LinkSetUp {
    dataList: any;
    Srno: number;
    LinkID: number;
    LinkName: string;
    IconName: string;
    url: string;
    Purpose: number;
    OrderNo: number;
    IsActive: boolean;
    CreatedOn: string;
    IsDefault: boolean;
    ParentMenuId: number;
    Flag: number;
}